Rename pilatus driver to io
