+++ src/ophyd_async/epics/adpilatus/__init__.py
@@ -1,11 +1,11 @@
 from ._pilatus import PilatusDetector, PilatusReadoutTime
 from ._pilatus_controller import PilatusController
-from ._pilatus_driver import PilatusDriver, PilatusTriggerMode
+from ._pilatus_io import PilatusDriverIO, PilatusTriggerMode
 
 __all__ = [
     "PilatusDetector",
     "PilatusReadoutTime",
     "PilatusController",
-    "PilatusDriver",
+    "PilatusDriverIO",
     "PilatusTriggerMode",
 ]
