+++ src/ophyd_async/epics/adpilatus/_pilatus_io.py
@@ -12,7 +12,7 @@ class PilatusTriggerMode(str, Enum):
     alignment = "Alignment"
 
 
-class PilatusDriver(adcore.ADBase):
+class PilatusDriverIO(adcore.ADBase):
     def __init__(self, prefix: str, name: str = "") -> None:
         self.trigger_mode = epics_signal_rw_rbv(
             PilatusTriggerMode, prefix + "TriggerMode"
